requirements.txt
eth_utils==4.1.1
web3==6.20.3
aiogram==3.18.0
aiohttp==3.9.5
cachetools==5.3.3
coincurve==19.0.1
cryptography==44.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
redis==5.2.1
asyncpg==0.30.0
pysha3==1.0.2
async_lru==2.0.4